training/requirements.txt
numpy>=1.24
psycopg2-binary>=2.9
pydantic>=2.0
pydantic-settings>=2.0
torch>=2.0
